updating files

+++ CSci5103/PA1/prodcon.c
@@ -1,23 +1,17 @@
 #include "prodcon.h"
 
 #define SHMKEY 2041
+#define SHMKEY2 2042
 struct sharedMemStruct{
   char * data;
   int * numChars;
 };
 
-void sigHandler(int sigNum){
-  switch(sigNum){
-    case SIGUSR1:
-      printf("handling SIGUSR1\n");
-      break;
-    case SIGUSR2:
-      printf("handling SIGUSR2\n");
-      break;
-    default:
-      printf("could not recognize signal\n");
-  }
-  
+void parentHandler(int sig){
+  printf("Parent caught signal\n");
+}
+void childHandler(int sig){
+  printf("Child caught signal\n");
 }
 
 void producerWriteText(char * filePath, pid_t childPID){
@@ -30,7 +24,7 @@ void producerWriteText(char * filePath, pid_t childPID){
   
   int sharedMemId, charsReadId;
   sharedMemId = shmget(SHMKEY, 1024, 0666 | IPC_CREAT);
-  charsReadId = shmget(SHMKEY, 4, 0666 | IPC_CREAT);
+  charsReadId = shmget(SHMKEY2, 4, 0666 | IPC_CREAT);
   if(sharedMemId == -1 || charsReadId == -1){
     perror("Error creating shared memory.\n");
     exit(1);
@@ -46,33 +40,67 @@ void producerWriteText(char * filePath, pid_t childPID){
   char *currentBlock = (char*)malloc(sizeof(char) * 1024);
   int charsRead = fread(currentBlock, 1, 1024, inputFile);
   //printf("chars read = %d\n", charsRead);
-  
+  int j = 0;
   while(charsRead > 0){
     int i;
-    
-    printf("waiting on signal SIGUSR2\n");
     
     for(i = 0; i < charsRead; i++){
       sharedMem.data[i] = currentBlock[i];
     }
-    *(sharedMem.numChars) = charsRead;
+    *(sharedMem.numChars) = (int)charsRead;
     printf("sending signal SIGUSR1\n");
+    sleep(1);
     kill(childPID, SIGUSR1);
-    sigpause(SIGUSR2);
-    
-    charsRead = fread(currentBlock, 1, 1024, inputFile); 
-   
+    printf("waiting on signal SIGUSR2\n");
+    //sigpause(SIGUSR2);
+    pause();
+    j = j+1;
+    charsRead = fread(currentBlock, 1, 1024, inputFile);
   }
   
+  *(sharedMem.numChars) = -1;
+  kill(childPID, SIGUSR1);
+  
   free(currentBlock);
+  free(sharedMem.data);
 }
 
 void consumerReadText(){
-
-  printf("waiting on signal SIGUSR1\n");
-  sigpause(SIGUSR1);
-  printf("sending signal SIGUSR2\n");
-  kill(getppid(), SIGUSR2);
+  FILE *outputFile = fopen("output", "w");
+  int sharedMemId, charsReadId;
+  sharedMemId = shmget(SHMKEY, 1024, 0666);
+  charsReadId = shmget(SHMKEY2, 4, 0666);
+  if(sharedMemId == -1 || charsReadId == -1){
+    perror("Error getting shared memory to consumer.\n");
+    exit(0);
+  }
+  
+  struct sharedMemStruct sharedMem;
+  sharedMem.data = shmat(sharedMemId, 0, 0);
+  sharedMem.numChars = shmat(charsReadId, 0, 0);
+  if(sharedMem.data == (char*) -1 || sharedMem.numChars == (int*) -1){
+    perror("Error attaching shared memory to consumer.\n");
+    exit(1);
+  }
+  int i = 0;
+  while(*sharedMem.numChars != -1){
+    printf("waiting on signal SIGUSR1\n");
+    
+    
+    //sigpause(SIGUSR1);
+    pause();
+    
+    printf("consumer recieved = \"%s\"\n", sharedMem.data);
+    printf("consumer recieved numChars = %d\"\n", *sharedMem.numChars);
+    
+    printf("sending signal SIGUSR2\n");
+    sleep(1);
+    kill(getppid(), SIGUSR2);
+    ++i;
+    if(i >= 500){
+      exit(0);
+    }
+  }
   printf("exiting...\n");
   exit(0);
   
@@ -82,19 +110,24 @@ int main(int argc, char ** argv){
   if(argc != 2){
     printf("user must input only the filename to be transfered.\n");
     exit(1);
-  } 
-
-  signal(SIGUSR1, sigHandler);
-  signal(SIGUSR2, sigHandler);
+  }
+  
+  static struct sigaction pAction, cAction;
+  pAction.sa_handler = parentHandler;
+  sigaction(SIGUSR1, &pAction, NULL);
   char* filePath = argv[1];
+  int status;
   
-  pid_t childPid = fork();  //fork a child process
+  pid_t childPid = fork(); //fork a child process
   
-  if(childPid == 0){  //child process goes to read function
+  if(childPid == 0){ //child process goes to read function
+    cAction.sa_handler = childHandler;
+    sigaction(SIGUSR1, &cAction, NULL);
     consumerReadText();
   }
   else if(childPid > 0){ //parent process
     producerWriteText(filePath, childPid);
+    wait(&status);
   }
   else{
     printf("fork error, exiting...\n");
@@ -102,6 +135,3 @@ int main(int argc, char ** argv){
   }
  return 1;
 }
-
-
-
